requirements.txt
scikit-learn
streamlit
streamlit-drawable-canvas
numpy
Pillow
matplotlib
seaborn